src/inpainting/cloud_function/requirements.txt
azure-functions
fastapi
pillow